Move make_redis_client_name and parse_redis_client_name to karton.core.backend.direct as it is Redis-related code
