+++ karton/core/backend/direct.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import time
+import urllib.parse
 import warnings
 from collections import defaultdict, namedtuple
 from typing import IO, Any, Dict, Iterable, Iterator, List, Optional, Set, Tuple, Union
@@ -23,14 +24,7 @@ from karton.core.resource import LocalResource, RemoteResource
 from karton.core.task import Task, TaskPriority, TaskState
 from karton.core.utils import chunks, chunks_iter
 
-from .base import (
-    KartonBackendProtocol,
-    KartonBind,
-    KartonMetrics,
-    KartonServiceInfo,
-    make_redis_client_name,
-    parse_redis_client_name,
-)
+from .base import KartonBackendProtocol, KartonBind, KartonMetrics, KartonServiceInfo
 
 KARTON_TASKS_QUEUE = "karton.tasks"
 KARTON_OPERATIONS_QUEUE = "karton.operations"
@@ -41,6 +35,28 @@ KARTON_OUTPUTS_NAMESPACE = "karton.outputs"
 
 KartonOutputs = namedtuple("KartonOutputs", ["identity", "outputs"])
 logger = logging.getLogger(__name__)
+
+
+def make_redis_client_name(service_info: KartonServiceInfo) -> str:
+    params = {
+        "karton_version": service_info.karton_version,
+    }
+    if service_info.service_version is not None:
+        params.update({"service_version": service_info.service_version})
+    return f"{service_info.identity}?{urllib.parse.urlencode(params)}"
+
+
+def parse_redis_client_name(client_name: str) -> KartonServiceInfo:
+    identity, params_string = client_name.split("?", 1)
+    # Filter out unknown params to not get crashed by future extensions
+    params = dict(urllib.parse.parse_qsl(params_string))
+    karton_version = params.get("karton_version", "")
+    service_version = params.get("service_version")
+    return KartonServiceInfo(
+        identity=identity,
+        karton_version=karton_version,
+        service_version=service_version,
+    )
 
 
 class KartonBackendBase:
